fix(worker): fixes timeserieses and data handling

+++ worker_v2_impl_go/compute/types.go
@@ -1,6 +1,6 @@
 package compute
 
 type resolution struct {
-	width  float64
-	height float64
+	width  int64
+	height int64
 }
